fix circular dependency
This moves code that is only used by c9 for the startup quickstart webview.
Maybe we can move this elsewhere.
